+++ src/DataProtection/DataProtection.Autorest/custom/Cmdlets/Platform/Vault/Set-AzDataProtectionMSIPermission.ps1
@@ -16,6 +16,7 @@ function Get-VaultIdentity {
         Write-Host "Using Vault UAMI with ARMId: $UserAssignedIdentityARMId with Principal ID: $vaultIdentity"
     } else {
         $vaultIdentity = $vault.Identity.PrincipalId
+        if (-not $vaultIdentity) { $vaultIdentity = $vault.IdentityPrincipalId }
         Write-Host "Using system-assigned identity with Principal ID: $vaultIdentity"
     }